Moved common methods in Utilities and added readme method
