+++ examples/workspace_example.py
@@ -902,16 +902,35 @@ class WorkspaceManager:
             print("   🗑️  Successfully deleted policy")
 
         except Exception as e:
-            print(f"   ❌ Data retention policy operations failed: {e}")
-            print("   💡 This may be due to:")
-            print(
-                "      • Insufficient permissions for data retention policy management"
-            )
-            print("      • Terraform Enterprise license requirements")
-            print("      • Network connectivity issues")
-            print("      • Workspace not found or inaccessible")
-            print("      • Organization-level policy restrictions")
-            print("      • Feature not available in Terraform Cloud")
+            error_msg = str(e).lower()
+            if "not found" in error_msg:
+                print(f"   ⚠️  Data retention policy feature not available: {e}")
+                print(
+                    "\n   💡 IMPORTANT: Data retention policies are a Terraform Enterprise feature"
+                )
+                print(
+                    "   📋 This feature is NOT available in Terraform Cloud (app.terraform.io)"
+                )
+                print("   🏢 To use data retention policies, you need:")
+                print("      • Terraform Enterprise (self-hosted)")
+                print("      • Terraform Business tier or higher")
+                print("      • Admin permissions on the organization")
+                print(
+                    "\n   ✅ This is expected behavior when running against Terraform Cloud"
+                )
+                print(
+                    "   📝 The implementation is correct and will work with Terraform Enterprise"
+                )
+            else:
+                print(f"   ❌ Data retention policy operations failed: {e}")
+                print("   💡 This may be due to:")
+                print(
+                    "      • Insufficient permissions for data retention policy management"
+                )
+                print("      • Terraform Enterprise license requirements")
+                print("      • Network connectivity issues")
+                print("      • Workspace not found or inaccessible")
+                print("      • Organization-level policy restrictions")
 
     def demo_delete_operations(
         self, organization: str, workspace_name: str, workspace_id: str
